Changes to be able to pass tests for erc20

+++ packages/nouns-contracts/contracts/interfaces/INounsAuctionHouse.sol
@@ -32,8 +32,7 @@ interface INounsAuctionHouse {
         // Whether or not the auction has been settled
         bool settled;
         // What was last bid
-        uint8 lastBidType;
-
+        uint256 lastBidType;
 
     }
